% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSARseq_version03.R
\name{DGEs}
\alias{DGEs}
\title{Extracting covariate information including adjustment factors from non-negative matrix factorization}
\usage{
DGEs(PSARobject)
}
\arguments{
\item{PSARObj}{An input object: output of PSAR() function}
}
\value{
\item{dgeset}{a covariate file}
}
\description{
Extracting covariate information including adjustment factors from non-negative matrix factorization
}
\details{
See the referenced paper.
}
\examples{
cnts <- matrix(rnbinom(1000, mu=10, size=1), nrow=100)
grps <- factor(rep(c("a", "b"), each=5))
res <- PSAR(cnts, grps, n.factor=1)
head(FACTORs(res))
}
\references{
Jungsoo Gim and Christoph Lange, PSARseq (2017)
}
\author{
Jungsoo Gim
}
