% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSARseq_version01.R
\name{PSAR}
\alias{PSAR}
\title{One-step function performing differential expression analysis adjusting population stratification}
\usage{
PSAR(Obj, grps, n.factor = 2)
}
\arguments{
\item{Obj}{An input object: SeqExpressionSet (EDASeq), DGEList (edgeR) or raw count in "matrix" class are available}

\item{grps}{A vector indicating the group of interest for DE test}

\item{n.factor}{The number of factors in non-negative factorization (default = 2)}
}
\value{
\item{resOut}{a list of including two variables: dgeset (DGEList) and fit (glmFit)}
}
\description{
One-step function performing differential expression analysis adjusting population stratification
}
\details{
See the referenced paper.
}
\examples{
(you example code, each line start with #')
}
\references{
Jungsoo Gim and Christoph Lange, PSARseq (2017)
}
\author{
Jungsoo Gim
}
